refactored project and added cv directory
