+++ utils/cv-directory/cv1.txt
@@ -1,0 +1,1 @@
+My name is Raul and I am a Computer Science student at Technical University of Cluj-Napoca, I have been working with Java for the past year and I have strong understanding of the Object Oriented Programming concepts and rules.
