+++ utils/cv-directory/cv4.txt
@@ -1,0 +1,5 @@
+Skills in:
+- Java for 3+ years
+- Open and willing to learn new technologies
+- Strong Java Core knowledge
+- Good understanding of design patterns
